Basic Overlay Working
The overlay I tried to emulate/incorporate from that GitHub link is
finally working. It needs a lot of cleaning up though, both source and
CSS. Currently, the "OPEN OVERLAY" button will actually toggle the
overlay. The X in the top right is optional for closing the overlay, at
the moment.