MapApp.js now has Leaflet.Draw
Included a block of code at the bottom of init() to make the appropriate
Leaflet.Draw control and put it on the map. Currently we can just draw
things on the map and edit them. I will need to research/work further to
actually start doing things with Leaflet.Draw now.