Locator Working
We now have a little locator button in the bottom right. It looks like a
standard leaflet marker. When clicked it will zoom in on the user's
location. I'd like to take this and start computing nearby stuff based
on location and perhaps put said info in sidebar to be brought up as
needed.