First Working LRM
So the LRM works right now as a basic proof-of-concept. By default, on
load, it puts a pin near SEL and one near the Field Museum. In the top
right are the directions. These directions update on the "drop" when you
drag & drop a pin. The map also zooms as needed to a bounding box
encompassing both pins and the control (with directions) updates and
expands as needed.

I'm going to need to tweak the markers and the control, etc.