Custom Icons for Potholes and Street Lights
It was quick and dirty, but potholes and street lights now each
respectively have custom icons. Quick and dirty because it's a very
un-graceful series of *if* statements that distinguish our data sets. A
switch/case or some other helper function should probably be used for
the final version for both a)code cleanliness and b)efficiency.