Crimes Update Properly Now
Our new crime markers are now not only created properly, but updated
properly too. Also, fixed an issue in the refreshCrimeData function that
was calling a nonexistent function, "getById", so I replaced the
function "getId" with "getById". The rectangle cursor continues to
behave as expected. Of course, area selection still suffers and
definitely can't make sense of multiple boxes.