Data as Markers!
I have moved my work work for custom markers here to master now.
Currently the CTA data has a placeholder marker and no popups are bound
to CTA data. This will all be changed later when I have the chance. At
least for now, it all works and it's just a series of cosmetic tweaks
that needs to be made.