Popups and Custom Icon
Added a popup for CTA bus markers. The heading is parsed into one of 8
basic compass directions at the beginning of addCTAMarkers(). Then,
since a non-delayed bus returns undefined, I set the default text as a
joke there. We can leave it or replace it with something a little
more... professional later if we need to. Oh, and bus markers now have a
new dedicated icon; attribution text has been updated as needed.