COLORS FINALLY
Each type of marker now has its own color associated with it. There is a
"_old" and "_new" version of each marker SVG intended to be used to
highlight new data. Currently the map is set to use all new icons. The
old icons are the same color, just a few shades darker.